requirements.txt
asgiref==3.5.2
blinker==1.4
brotli==1.0.9
brotlicffi==1.0.9.2
cdecimal==2.3
closure_soy==20121221
ConfigParser==5.2.0
cryptography==37.0.2
cx_Oracle==8.3.0
Cython==0.29.30
dl==0.1.0
docutils==0.18.1
flask_azure_storage==0.2.1
flask_cdn==1.5.3
flask_s3==0.3.3
flask_script==2.0.6
genapi==0.0.8
glob2==0.7
HTMLParser==0.0.2
hypothesis==6.46.11
imageio==2.19.3
ipython==8.4.0
jnius==1.1.0
keyring==23.5.1
lockfile==0.12.2
lxml==4.9.0
monotonic==1.6
mtrand==0.1
mx==3.2
nose==1.3.7
numarray==1.5.1
Numeric==24.2
ordereddict==1.1
pgdb==0.0.11
pickle5==0.0.12
Pillow==9.1.1
protobuf==4.21.1
psutil==5.9.1
pyOpenSSL==22.0.0
pyScss==1.4.0
pysqlcipher3==1.0.4
pysqlite==2.8.3
pytest==7.1.2
pytz==2020.1
PyYAML==6.0
scipy==1.8.1
sets==0.3.2
simplejson==3.17.6
sqlcipher3==0.4.5
Sybase==0.35
tornado==6.1
typing_extensions==4.2.0
unicodedata2==14.0.0
watchdog==2.1.6
wincertstore==0.2.1
xmlrpclib==1.0.1
zope==5.5.1
